+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+2025-07-02 (v.0.4.0)
+--------------------
+- Added `waxmann_graph()` and `build_waxmann_graph()` for constructing Waxmann graphs with associated node coordinates.
+- Joint plotting of graphs and coordinates: `plot_graph()` method with graph and `Coordinates`, `plot_node_locations()` can take graphs to determine which (special) edges to draw.
+
 2025-06-09 (v.0.3.0)
 --------------------
 - Added an optional `rng` parameter for each of the randomized graph generation functions.